Extract Project Table helper
